refactor: move review_comment mutations under application_review_step
