+++ app/containers/Admin/ApplicationReview/ReviewSidebar.tsx
@@ -5,8 +5,8 @@ import {faCheck, faLock} from '@fortawesome/free-solid-svg-icons';
 import {graphql, createFragmentContainer, RelayProp} from 'react-relay';
 import {ReviewSidebar_applicationReviewStep} from '__generated__/ReviewSidebar_applicationReviewStep.graphql';
 import ReviewComment from 'components/Admin/ReviewComment';
-import updateReviewCommentMutation from 'mutations/application/updateReviewCommentMutation';
-import deleteReviewCommentMutation from 'mutations/application/deleteReviewCommentMutation';
+import updateReviewCommentMutation from 'mutations/application_review_step/updateReviewCommentMutation';
+import deleteReviewCommentMutation from 'mutations/application_review_step/deleteReviewCommentMutation';
 import {nowMoment} from 'functions/formatDates';
 import {capitalize} from 'lib/text-transforms';
